# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eden Life",
        "description": "Welcome to the Garden of Eden. it's new day, another opportunity to be a step closer to 'Paradise'!",
        "contact": {
            "email": "olusolagloryolamide@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/countries/{countryId}": {
            "get": {
                "tags": [
                    "countries"
                ],
                "summary": "Country of domicile by id",
                "description": "Get country by id",
                "operationId": "countriesById",
                "parameters": [
                    {
                        "name": "countryId",
                        "in": "path",
                        "description": "ID of country",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Country with given ID not found"
                    }
                }
            }
        },
...
        "/api/v1/gardenersByCountry/{countryId}": {
            "get": {
                "tags": [
                    "gardenersByCountry"
                ],
                "summary": "Country of domicile by id and gardeners per country and their respective numbers of  customers they have each",
                "description": "Get country by id and gardeners per country and their respective numbers of  customers they have each",
                "operationId": "countriesById",
                "parameters": [
                    {
                        "name": "countryId",
                        "in": "path",
                        "description": "ID of country",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Gardener with given ID not found"
                    }
                }
            }
        },
...
        "/api/v1/locations/{locationId}": {
            "get": {
                "tags": [
                    "locations"
                ],
                "summary": "Location Area by id and their respective customers",
                "description": "Get location area by id and their respective customers",
                "operationId": "locationsById",
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "path",
                        "description": "ID of location",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Location with given ID not found"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}